change second

+++ src/BmiServise.java
@@ -1,6 +1,6 @@
 public class BmiServise {
-    public double calculate(double weight, double hight) {
+    public int calculate(int weight, double hight) {
         double index = weight / (hight * hight);
-        return index;
+        return (int) index;
     }
 }
